Finished round two routing
This layout cuts through the middle of no planes and has more
predictable power paths (without needing Hyperlynx or something like
that). It also routes all signals but the LED-related ones on just the
top and bottom layers, as it is meant to be. It also nearly comforms to
the recommended layout for the regulator. There also aren't any ICs
without a full reference plane beneath them.

Basically, it's awesome.