Checked layout and reference designs, swapped crystal and RGB LED
ST has a list of suggested crystals which are known to work with the
STM32L0 family. I selected one of those and it is significantly
smaller in all dimensions.

The RGB LED had too high of a green voltage. I swapped it out for a
slightly more expensive one with the same footprint that had startlingly
low green and blue forward voltages.

For layout changes:
- Fixed regulator footprint silkscreen
- Fixed regulator layout to be closer to reference design
- Fixed inductor silkscreen
- Moved accelerometer downward to accomodate capacitors above
- Fixed accelerometer layout to have no traces beneath
- Fixed accelerometer silkscreen
- Fixed LED silkscreen
- General small silkscreen edits, such as shifting text around

I've also chosen parts and values for all resistors and capacitors.