Fixed LED ISR and buzzer duty cycle voltage droop
LED ISR: I forgot to reset the status register, so it would continuously
execute. I also found that I needed to up the refresh rate by a lot to
remove visible flicker. I suspect this stems from the fact that all
segments are blanked except the one currently being illuminated.

Buzzer duty cycle: The buzzer at 50% draws ~100mA average, with ~200mA
peak, according to the datasheet. On the voltage rail, there is an
extremely noticable droop, though it follows a roughly triangular curve,
likely due to the inductance of the buzzer coil (it is magnetic).
Reducing the duty cycle to roughly 2% completely eliminates the
resulting LED flicker at the expense of adding some high pitched
harmonics to the sound (it sounds a tad odd).