Added initial battery charge & usb detection, required bodge
The bodge was adding a 4.7K resistor between the base and emitter of Q1
and changing R4 to be 20K. The issue is that *something* is leaking 2V
onto the 5V line, which is enough to turn on Q1 through a 10K resistor.
I suspect the battery charger, but I guess it could be the diode I
specifically selected for a low leakage current. In any case, the
20K/4.7K voltage divider prevents the transistor from turning on until
at least ~4V is present on the line. Ideally it would have been exactly
4.5V, but my supply of surface mount resistors is a tad limited.