Fixed excessive sleep current consumption originating in the blue LEDs
Turns out, the blue LEDs don't have a reverse current specification and
they weren't designed to operate under reverse voltage. I missed that
little tidbit in the datasheet and now I know what they do when reverse
voltaged: They draw collectively 28mA or so through the microcontroller,
even when sleeping. The reason is that since the mux is disabled and the
hour pin is most likely set low (unless it happened to turn off when
muxing through the active hour segment), that current is drawn in
reverse through the LEDs. The software fix is to simply set the blue pin
high when we disable the LEDs. Current consumption in sleep mode drops
from ~30mA down to 0.45mA.

This does not fix the issue of excessive power usage when the display is
turned on. To fix that, I need new LEDs. However, these blue LEDs were
chosen because of their excessively low forward voltage and I have a
feeling that that is the reason they behave so strangely when reversed.