Fixed problems with NOZLP and specific packet sizes
In the event that a packet size was a multiple of the endpoint size, the
device would be unable to send or receive more than the first <endpoint
size> bytes. Additional logic was added to ensure that transfers are
only completed once the expected number of bytes were sent or received.
This still adds some additional peril for receivers which expect packets
longer than the endpoint size since it will no longer complete a receive
operation if the number of bytes received is a multiple of the endpoint
size (since there is no longer a way to tell if a transmission is
complete in that case).

Long story short, it won't choke on packets longer than the endpoint
size in NOZLP mode, but mistakes on the host side in sending packets
which are multiples of the endpoint size could lead to the device
getting out of sync with whatever is sending USB packets.