Increased poll rate to maximum and added progress bar to polling script
I believe that the data rate is being limited by the handshaking that
happens by enforcing a status report to be sent after every page
command. However, this does allow us to stop programming and determine
which page caused the error. Perhaps I need to rethink the status
reports so that I can remove the ordering dependency.